requirements.txt
# Azure RAGcelerator Dependencies
# Python 3.11+ required

# Core Azure SDKs
azure-identity>=1.15.0
azure-storage-blob>=12.19.0
azure-search-documents>=11.4.0
azure-functions>=1.17.0

# AI/ML
openai>=1.12.0
langchain>=0.1.0
langchain-text-splitters>=0.0.1

# PDF Processing
PyPDF2>=3.0.0
pdfplumber>=0.10.0

# Web UI
streamlit>=1.31.0

# Configuration & Utilities
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# Observability
opencensus-ext-azure>=1.1.0

# Development & Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
ruff>=0.2.0
black>=24.1.0

# Type checking (optional but recommended)
mypy>=1.8.0
types-PyPDF2>=3.0.0



